requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pandas==2.2.2